+++ apps/web/src/features/diagram/index.ts
@@ -2,6 +2,9 @@
 export { default as Canvas } from './components/Canvas';
 export { default as MemoryLayer } from './components/MemoryLayer';
 
+// Diagram UI wrappers (only the ones with store integration)
+export { CustomArrow, ContextMenu } from './wrappers';
+
 // Hooks
 export * from './hooks/useDiagramRunner';
 export * from './hooks/useDiagramActions';
