+++ apps/web/src/features/diagram/wrappers/index.ts
@@ -1,0 +1,12 @@
+// Diagram UI package wrappers with store integration
+export { CustomArrow } from '../components/CustomArrow';
+export { ContextMenu } from '../components/ContextMenu';
+
+// Direct exports from diagram-ui package (no store integration needed)
+export { FlowHandle, createHandleId, useKeyboardShortcuts, useContextMenu } from '@repo/diagram-ui';
+
+// Node components (re-export from diagram-ui)
+export { BaseNode, GenericNode } from '@repo/diagram-ui';
+
+// Re-export types
+export type { CustomArrowProps, ContextMenuProps, FlowHandleProps, BaseNodeProps, GenericNodeProps } from '@repo/diagram-ui';
