+++ apps/web/src/features/properties/wrappers/index.ts
@@ -1,0 +1,12 @@
+// Direct exports from properties-ui package (no store integration needed)
+export { Panel, Form, FormField, FormGrid, FormRow, TextAreaField, TextField, SelectField, InlineTextField, InlineSelectField } from '@repo/properties-ui';
+
+// Hooks
+export { usePropertyForm } from '../hooks/usePropertyForm';
+export { usePropertyPanel } from '../hooks/usePropertyPanel';
+
+// Re-export types
+export type {
+  PanelProps,
+  FormFieldProps
+} from '@repo/properties-ui';
